implement drag and drop
For brevity, only a single drag icon at a time is supported.

Some checks are removed from pointerfocus because
wlr_seat_pointer_notify_motion() needs to be called even if a surface
isn't focused for dragging to work, and because they're redundant since
they are already performed by wlroots, and in fact they have been
removed from tinywl too in the wlroots commit f2f3df9.

dragdestroy doesn't restore pointer focus with motionnotify(0) because
it makes mouse input stop working in Firefox for some reason (but not in
Chromium). It also doesn't try to focus any layer surface or unmanaged
client since I can't think of any that supports drag and drop, and it
would require keeping track of the focused layer surface and complicate
the code.